+++ blobs/asset_test.go
@@ -12,6 +12,7 @@ func TestCanSeeBaseZipAsset(t *testing.T) {
 	must_be, wont_be := hamlet.Specifications(t)
 
 	must_be.Panic(func() { blobs.MustAsset("assets/missing.zip") })
+	wont_be.Panic(func() { blobs.MustAsset("assets/templates.yaml") })
 	wont_be.Panic(func() { blobs.MustAsset("assets/standard.zip") })
 	wont_be.Panic(func() { blobs.MustAsset("assets/python.zip") })
 	wont_be.Panic(func() { blobs.MustAsset("assets/settings.yaml") })
@@ -27,7 +28,7 @@ func TestCanSeeBaseZipAsset(t *testing.T) {
 func TestCanGetTemplateNamesThruOperations(t *testing.T) {
 	must_be, wont_be := hamlet.Specifications(t)
 
-	assets := operations.ListTemplates()
+	assets := operations.ListTemplates(true)
 	wont_be.Nil(assets)
 	must_be.True(len(assets) == 3)
 	must_be.Equal([]string{"extended", "python", "standard"}, assets)